Enh: remove CodeDom dependency from runtime
