+++ XObjectsCode/Configuration/ConfigurationProvider.cs
