+++ XObjectsCode/Configuration/LinqToXsdConfiguration.Custom.cs
