+++ XObjectsCode/Configuration/NamespaceEqualityValueComparer.cs
