+++ XObjectsCode/Src/CodeMemberFieldEqualityComparer.cs
