+++ XObjectsCode/Src/XObjectsCoreGenerator.cs
