+++ XObjectsCore/API/ChoiceContentModelEntity.cs
