+++ XObjectsCore/API/ContentModel.cs
