+++ XObjectsCore/API/FsmXObjects.cs
