+++ XObjectsCore/API/Interfaces.cs
