+++ XObjectsCore/API/NamedContentModelEntity.cs
