+++ XObjectsCore/API/SchemaAwareContentModelEntity.cs
