+++ XObjectsCore/API/SequenceContentModelEntity.cs
