+++ XObjectsCore/API/SubstitutedContentModelEntity.cs
