+++ XObjectsCore/API/SubstitutionMembersList.cs
