+++ XObjectsCore/API/TypeManager.cs
