+++ XObjectsCore/API/XList.cs
