+++ XObjectsCore/API/XListVisualizable.cs
