+++ XObjectsCore/API/XNamespaceResolver.cs
