+++ XObjectsCore/API/XObjects.cs
