+++ XObjectsCore/API/XObjectsQueryExtension.cs
