+++ XObjectsCore/API/XSimpleList.cs
