+++ XObjectsCore/API/XTypedElementEqualityComparer.cs
