+++ XObjectsCore/API/XTypedList.cs
