+++ XObjectsCore/API/XTypedServices.cs
