+++ XObjectsCore/API/XTypedSubstitutedList.cs
