+++ XObjectsCore/Constants.cs
