+++ XObjectsCore/Enums.cs
