+++ XObjectsCore/FSM.cs
