+++ XObjectsCore/NamespaceList.cs
