+++ XObjectsCore/XObjectException.cs
