+++ XObjectsCore/XmlComplianceUtil.cs
